fix: remove circular dependencies
