+++ .changeset/rich-grapes-cheat.md
@@ -1,0 +1,6 @@
+---
+"marko": patch
+"@marko/translator-default": patch
+---
+
+Removes circular dependencies from the Marko runtime.
