--- packages/marko/src/runtime/components/init-components/index.js
@@ -1,1 +1,0 @@
-// The server-side implementation of this module is intentionally empty
