--- packages/marko/src/runtime/components/init-components/package.json
@@ -1,11 +1,0 @@
-{
-  "main": "./index.js",
-  "browser": "./index-browser.js",
-  "exports": {
-    ".": {
-      "worker": "./index.js",
-      "browser": "./index-browser.js",
-      "default": "./index.js"
-    }
-  }
-}
